[Refactor] 장소 관련 모든 아이템 누를때 보내는 값 변경
Place Id : Long -> Int로 타입 전부 변경

+++ app/src/main/java/place/pic/data/remote/response/PlaceItemResponse.kt
@@ -17,7 +17,7 @@ data class PlaceResponse(
 }
 
 data class PlaceItemResponse(
-    val placeIdx: Long,
+    val placeIdx: Int,
     val placeName: String,
     val placeCreatedAt: Long,
     val tag: List<TagResponse>,
@@ -39,7 +39,7 @@ data class PlaceItemResponse(
 
     fun toPlaceGridItem() = PlaceGridItem(
         imageUrl = imageUrl.getOrElse(0) { "" },
-        placeIdx = placeIdx.toInt(),
+        placeIdx = placeIdx,
         placeName = placeName,
         likeCount = likeCount,
         subwayName = subway.map { it.subwayName }
